add checkout metadata and invoice number
