+++ excise/plugin.py
@@ -1,5 +1,6 @@
 import json
 import logging
+from dataclasses import asdict
 from decimal import Decimal
 from typing import TYPE_CHECKING, Any, Iterable, Optional
 from urllib.parse import urljoin
@@ -9,31 +10,34 @@ import opentracing.tags
 from django.core.exceptions import ValidationError
 from prices import Money, TaxedMoney
 
-from ....checkout.models import Checkout
-from ....core.prices import quantize_price
-from ....core.taxes import TaxError, zero_taxed_money
-from ....discount import DiscountInfo
-from ...base_plugin import ConfigurationTypeField
-from ...error_codes import PluginErrorCode
-from .. import _validate_checkout
-from ..plugin import AvataxPlugin
-from . import (
+from saleor.checkout.models import Checkout
+from saleor.core.prices import quantize_price
+from saleor.core.taxes import TaxError, zero_taxed_money
+from saleor.discount import DiscountInfo
+from saleor.plugins.base_plugin import ConfigurationTypeField
+from saleor.plugins.error_codes import PluginErrorCode
+from saleor.plugins.avatax import _validate_checkout
+from saleor.plugins.avatax.plugin import AvataxPlugin
+from .utils import (
     api_get_request,
     api_post_request,
     generate_request_data_from_checkout,
     get_api_url,
     get_checkout_tax_data,
-    get_order_tax_data,
+    get_order_request_data,
+    TRANSACTION_TYPE,
+    process_checkout_metadata
 )
+from .tasks import api_post_request_task
 
 if TYPE_CHECKING:
     # flake8: noqa
-    from ....account.models import Address
-    from ....channel.models import Channel
-    from ....checkout import CheckoutLineInfo
-    from ....checkout.models import CheckoutLine
-    from ....order.models import Order
-    from ...models import PluginConfiguration
+    from saleor.account.models import Address
+    from saleor.channel.models import Channel
+    from saleor.checkout import CheckoutLineInfo
+    from saleor.checkout.models import CheckoutLine
+    from saleor.order.models import Order
+    from saleor.plugins.models import PluginConfiguration
 
 logger = logging.getLogger(__name__)
 
@@ -123,8 +127,7 @@ class AvataxExcisePlugin(AvataxPlugin):
     def calculate_checkout_total(
         self,
         checkout: "Checkout",
-        lines: Iterable["CheckoutLineInfo"],
-        address: Optional["Address"],
+        lines: Iterable["CheckoutLine"],
         discounts: Iterable[DiscountInfo],
         previous_value: TaxedMoney,
     ) -> TaxedMoney:
@@ -133,7 +136,7 @@ class AvataxExcisePlugin(AvataxPlugin):
             return previous_value
         checkout_total = previous_value
 
-        if not _validate_checkout(checkout, [line_info.line for line_info in lines]):
+        if not _validate_checkout(checkout, lines):
             logger.debug("Checkout Invalid in Calculate Checkout Total")
             return checkout_total
 
@@ -146,20 +149,15 @@ class AvataxExcisePlugin(AvataxPlugin):
 
         currency = checkout.currency
 
-        # store itemized tax information in Checkout metadata for optional display on the frontend
-        # if there are no taxes, itemized taxes = []
-        tax_item = {"itemized_taxes": json.dumps(response["TransactionTaxes"])}
-        checkout_obj = Checkout.objects.filter(token=checkout.token).first()
-        if checkout_obj:
-            checkout_obj.store_value_in_metadata(items=tax_item)
-            checkout_obj.save()
-
         tax = Money(Decimal(response.get("TotalTaxAmount", 0.0)), currency)
         net = checkout_total.net
         total_gross = net + tax
         taxed_total = quantize_price(TaxedMoney(net=net, gross=total_gross), currency)
         total = self._append_prices_of_not_taxed_lines(
-            taxed_total, lines, checkout.channel, discounts
+            # taxed_total, lines, checkout.channel, discounts
+            taxed_total,
+            lines,
+            discounts,
         )
 
         voucher_value = checkout.discount
@@ -171,8 +169,7 @@ class AvataxExcisePlugin(AvataxPlugin):
     def calculate_checkout_subtotal(
         self,
         checkout: "Checkout",
-        lines: Iterable["CheckoutLineInfo"],
-        address: Optional["Address"],
+        lines: Iterable["CheckoutLine"],
         discounts: Iterable[DiscountInfo],
         previous_value: TaxedMoney,
     ) -> TaxedMoney:
@@ -181,8 +178,7 @@ class AvataxExcisePlugin(AvataxPlugin):
     def calculate_checkout_shipping(
         self,
         checkout: "Checkout",
-        lines: Iterable["CheckoutLineInfo"],
-        address: Optional["Address"],
+        lines: Iterable["CheckoutLine"],
         discounts: Iterable[DiscountInfo],
         previous_value: TaxedMoney,
     ) -> TaxedMoney:
@@ -198,13 +194,12 @@ class AvataxExcisePlugin(AvataxPlugin):
 
         Raise an error when can't receive taxes.
         """
+
         if self._skip_plugin(previous_value):
             return previous_value
 
         data = generate_request_data_from_checkout(
-            checkout,
-            transaction_type="RETAIL",
-            discounts=discounts,
+            checkout, transaction_type=TRANSACTION_TYPE, discounts=discounts,
         )
         if not data.TransactionLines:
             return previous_value
@@ -238,58 +233,51 @@ class AvataxExcisePlugin(AvataxPlugin):
         return previous_value
 
     def order_created(self, order: "Order", previous_value: Any) -> Any:
-
-        # call the create transactions (similar flow as calculate checkout total)
-        response = get_order_tax_data(order, self.config)
-
-        # TODO handle error if get_order_tax_data fails
-
-        transaction_id = response.get("UserTranId")
-
-        if self.config.autocommit:
-            # call the commit api with the UserTranId
-            commit_url = urljoin(
-                get_api_url(self.config.use_sandbox),
-                f"AvaTaxExcise/transactions/{transaction_id}/commit",
-            )
-            commit_response = api_post_request(commit_url, None, self.config)
-
-            # TODO do something if commit fails
+        if not self.active:
+            return previous_value
+        request_data = get_order_request_data(order)
+        # TODO Needed for commit
+        # transaction_id = response.get("UserTranId")
+        transaction_url = urljoin(
+            get_api_url(self.config.use_sandbox), "AvaTaxExcise/transactions/create",
+        )
+        api_post_request_task.delay(
+            transaction_url, asdict(request_data), asdict(self.config), order.id
+        )
 
         return previous_value
 
     def calculate_checkout_line_total(
         self,
-        checkout: "Checkout",
-        checkout_line_info: "CheckoutLineInfo",
-        address: Optional["Address"],
-        channel: "Channel",
-        discounts: Iterable["DiscountInfo"],
+        checkout_line: "CheckoutLine",
+        discounts: Iterable[DiscountInfo],
         previous_value: TaxedMoney,
     ) -> TaxedMoney:
         if self._skip_plugin(previous_value):
-            # logger.debug("Skip plugin %s", previous_value)
             return previous_value
 
         base_total = previous_value
-        if not checkout_line_info.product.charge_taxes:
-            # logger.debug("Charge taxes is false for this item %s")
+        if not checkout_line.variant.product.charge_taxes:
             return base_total
 
-        if not _validate_checkout(checkout, [checkout_line_info.line]):
-            # logger.debug("Checkout invalid %s")
+        checkout = checkout_line.checkout
+
+        if not _validate_checkout(checkout, [checkout_line]):
             return base_total
 
         taxes_data = get_checkout_tax_data(checkout, discounts, self.config)
 
         if not taxes_data or "Error" in taxes_data["Status"]:
-            logger.debug("Error in tax response %s")
             return base_total
+
+        # store itemized tax information in Checkout metadata for optional display on the frontend
+        tax_meta = json.dumps(taxes_data["TransactionTaxes"])
+        process_checkout_metadata(tax_meta, checkout.token)
 
         line_tax_total = Decimal(0)
 
         for line in taxes_data.get("TransactionTaxes", []):
-            if line.get("InvoiceLine") == checkout_line_info.line.id:
+            if line.get("InvoiceLine") == checkout_line.id:
                 line_tax_total += Decimal(line.get("TaxAmount", 0.0))
 
         if not line_tax_total > 0:
