+++ excise/tests/cassettes/test_order_created.yaml
@@ -1,7 +1,7 @@
 interactions:
 - request:
     body: '{"EffectiveDate": "02/24/21", "InvoiceDate": "02/24/21", "TitleTransferCode":
-      "DEST", "TransactionType": "RETAIL", "TransactionLines": [{"InvoiceLine": 72,
+      "DEST", "TransactionType": "DIRECT", "TransactionLines": [{"InvoiceLine": 72,
       "ProductCode": "202015500", "UnitPrice": "170.000", "UnitOfMeasure": "PAC",
       "BilledUnits": "3", "LineAmount": "12.300", "AlternateUnitPrice": "1.000", "TaxIncluded":
       true, "UnitQuantity": null, "UnitQuantityUnitOfMeasure": "EA", "DestinationCountryCode":
