+++ excise/tests/cassettes/test_preprocess_order_creation.yaml
@@ -1,7 +1,7 @@
 interactions:
 - request:
     body: '{"EffectiveDate": "02/18/21", "InvoiceDate": "02/18/21", "TitleTransferCode":
-      "DEST", "TransactionType": "RETAIL", "TransactionLines": [{"InvoiceLine": 9,
+      "DEST", "TransactionType": "DIRECT", "TransactionLines": [{"InvoiceLine": 9,
       "ProductCode": "123", "UnitPrice": "10.000", "UnitOfMeasure": "PAC", "BilledUnits":
       "3", "LineAmount": "15.00", "AlternateUnitPrice": "1.000", "TaxIncluded": true,
       "UnitQuantity": null, "UnitQuantityUnitOfMeasure": "EA", "DestinationCountryCode":
