+++ excise/tests/test_avatax_excise.py
@@ -1,6 +1,7 @@
 from decimal import Decimal
 from json import JSONDecodeError
 from unittest.mock import Mock, patch
+from urllib.parse import urljoin
 
 import pytest
 from django.core.exceptions import ValidationError
@@ -9,40 +10,41 @@ from django.db import connection
 from django.test import override_settings
 from prices import Money, TaxedMoney
 from requests import RequestException
+from dataclasses import asdict
 
-from saleor.shipping.models import ShippingMethodChannelListing
 
-from .....account.models import Address
-from .....checkout import CheckoutLineInfo
-from .....checkout.models import CheckoutLine
-from .....checkout.utils import add_variant_to_checkout, fetch_checkout_lines
-from .....core.prices import quantize_price
-from .....core.taxes import TaxError
-from .....product.models import ProductType, ProductVariant
-from .....warehouse.models import Warehouse
+# from saleor.shipping.models import ShippingMethodChannelListing
+
+from saleor.account.models import Address
+
+# from saleor.checkout import CheckoutLineInfo
+from saleor.checkout.models import CheckoutLine, Checkout
+from saleor.checkout.utils import add_variant_to_checkout
+from ..compat import fetch_checkout_lines
+from saleor.core.prices import quantize_price
+from saleor.core.taxes import TaxError
+from saleor.product.models import ProductType, ProductVariant
+from saleor.warehouse.models import Warehouse
 from ....manager import get_plugins_manager
 from ....models import PluginConfiguration
 from ... import AvataxConfiguration
-from .. import api_post_request, get_metadata_key
+from ..utils import api_post_request, get_metadata_key, get_api_url, get_order_request_data
 from ..plugin import AvataxExcisePlugin
 
 
 @pytest.fixture
 def plugin_configuration(db):
     def set_configuration(
-        username="test",
-        password="test",
-        sandbox=True,
-        company_id="test",
+        username="test", password="test", sandbox=True, company_id="test",
     ):
         data = {
             "active": True,
             "name": AvataxExcisePlugin.PLUGIN_NAME,
             "configuration": [
-                {"name": "Username or account", "value": username},
-                {"name": "Password or license", "value": password},
-                {"name": "Use sandbox", "value": sandbox},
-                {"name": "Company name", "value": company_id},
+                {"name": "Username or account", "value": "test"},
+                {"name": "Password or license", "value": "test"},
+                {"name": "Use sandbox", "value": True},
+                {"name": "Company name", "value": "test"},
                 {"name": "Autocommit", "value": False},
             ],
         }
@@ -172,9 +174,7 @@ def test_save_plugin_configuration_authentication_failed(
 @pytest.mark.vcr
 @pytest.mark.parametrize(
     "with_discount, expected_net, expected_gross, taxes_in_prices",
-    [
-        (False, "30.00", "32.29", True),
-    ],
+    [(False, "30.00", "32.29", True),],
 )
 @override_settings(PLUGINS=["saleor.plugins.avatax.excise.plugin.AvataxExcisePlugin"])
 def test_calculate_checkout_line_total(
@@ -206,36 +206,81 @@ def test_calculate_checkout_line_total(
     product.save()
     product.product_type.save()
     discounts = [discount_info] if with_discount else None
-    channel = checkout_with_item.channel
-    channel_listing = line.variant.channel_listings.get(channel=channel)
+    # channel = checkout_with_item.channel
+    # channel_listing = line.variant.channel_listings.get(channel=channel)
 
-    checkout_line_info = CheckoutLineInfo(
-        line=line,
-        variant=line.variant,
-        channel_listing=channel_listing,
-        product=line.variant.product,
-        collections=[],
-    )
+    # checkout_line_info = CheckoutLineInfo(
+    #     line=line,
+    #     variant=line.variant,
+    #     channel_listing=channel_listing,
+    #     product=line.variant.product,
+    #     collections=[],
+    # )
+    line = checkout_with_item.lines.first()
 
-    total = manager.calculate_checkout_line_total(
-        checkout_with_item,
-        checkout_line_info,
-        checkout_with_item.shipping_address,
-        channel,
-        discounts,
-    )
+    total = manager.calculate_checkout_line_total(line, discounts,)
     total = quantize_price(total, total.currency)
     assert total == TaxedMoney(
         net=Money(expected_net, "USD"), gross=Money(expected_gross, "USD")
     )
 
+@pytest.mark.vcr
+@pytest.mark.parametrize(
+    "with_discount, expected_net, expected_gross, taxes_in_prices",
+    [(False, "30.00", "32.29", True),],
+)
+@override_settings(PLUGINS=["saleor.plugins.avatax.excise.plugin.AvataxExcisePlugin"])
+def test_calculate_checkout_line_total_metadata(
+    reset_sequences,  # pylint: disable=unused-argument
+    with_discount,
+    expected_net,
+    expected_gross,
+    taxes_in_prices,
+    discount_info,
+    checkout_with_item,
+    address_usa_tx,
+    address_usa,
+    site_settings,
+    shipping_zone,
+    plugin_configuration,
+):
+    plugin_configuration()
+    manager = get_plugins_manager()
+
+    checkout_with_item.shipping_address = address_usa_tx
+    checkout_with_item.shipping_method = shipping_zone.shipping_methods.get()
+    checkout_with_item.save()
+    site_settings.company_address = address_usa
+    site_settings.include_taxes_in_prices = taxes_in_prices
+    site_settings.save()
+    line = checkout_with_item.lines.first()
+    product = line.variant.product
+    product.metadata = {}  # TODO consider adding ATE fields here
+    product.save()
+    product.product_type.save()
+    discounts = [discount_info] if with_discount else None
+    # channel = checkout_with_item.channel
+    # channel_listing = line.variant.channel_listings.get(channel=channel)
+
+    # checkout_line_info = CheckoutLineInfo(
+    #     line=line,
+    #     variant=line.variant,
+    #     channel_listing=channel_listing,
+    #     product=line.variant.product,
+    #     collections=[],
+    # )
+    line = checkout_with_item.lines.first()
+
+    manager.calculate_checkout_line_total(line, discounts,)
+
+    checkout = Checkout.objects.filter(token=line.checkout.token).first()
+    assert checkout.metadata[get_metadata_key("itemized_taxes")] == '[{"TransactionTaxAmounts": [], "SequenceId": 1, "TransactionLine": 1, "InvoiceLine": 1, "CountryCode": "USA", "Jurisdiction": "TX", "LocalJurisdiction": "48", "ProductCategory": 0.0, "TaxingLevel": "STA", "TaxType": "S", "RateType": "G", "RateSubtype": "NONE", "CalculationTypeInd": "P", "TaxRate": 0.0625, "TaxQuantity": 0.0, "TaxAmount": 1.73, "TaxExemptionInd": "N", "SalesTaxBaseAmount": 27.71, "LicenseNumber": "", "RateDescription": "TX STATE TAX - TEXAS", "Currency": "USD", "SubtotalInd": "C", "StatusCode": "ACTIVE", "QuantityInd": "B"}, {"TransactionTaxAmounts": [], "SequenceId": 2, "TransactionLine": 1, "InvoiceLine": 1, "CountryCode": "USA", "Jurisdiction": "TX", "LocalJurisdiction": "05000", "ProductCategory": 0.0, "TaxingLevel": "CIT", "TaxType": "S", "RateType": "G", "RateSubtype": "NONE", "CalculationTypeInd": "P", "TaxRate": 0.01, "TaxQuantity": 0.0, "TaxAmount": 0.28, "TaxExemptionInd": "N", "SalesTaxBaseAmount": 27.71, "LicenseNumber": "", "RateDescription": "TX CITY TAX - AUSTIN", "Currency": "USD", "SubtotalInd": "C", "StatusCode": "ACTIVE", "QuantityInd": "B"}, {"TransactionTaxAmounts": [], "SequenceId": 3, "TransactionLine": 1, "InvoiceLine": 1, "CountryCode": "USA", "Jurisdiction": "TX", "LocalJurisdiction": "6000814", "ProductCategory": 0.0, "TaxingLevel": "STJ", "TaxType": "S", "RateType": "G", "RateSubtype": "NONE", "CalculationTypeInd": "P", "TaxRate": 0.01, "TaxQuantity": 0.0, "TaxAmount": 0.28, "TaxExemptionInd": "N", "SalesTaxBaseAmount": 27.71, "LicenseNumber": "", "RateDescription": "TX SPECIAL TAX - AUSTIN MTA TRANSIT", "Currency": "USD", "SubtotalInd": "C", "StatusCode": "ACTIVE", "QuantityInd": "B"}]'
+
 
 @pytest.mark.vcr
 @pytest.mark.parametrize(
     "with_discount, expected_net, expected_gross, voucher_amount, taxes_in_prices",
-    [
-        (False, "43.98", "48.95", "0.0", False),
-    ],
+    [(False, "43.98", "48.95", "0.0", False),],
 )
 @override_settings(PLUGINS=["saleor.plugins.avatax.excise.plugin.AvataxExcisePlugin"])
 def test_calculate_checkout_total(
@@ -290,7 +335,7 @@ def test_calculate_checkout_total(
     discounts = [discount_info] if with_discount else None
     lines = fetch_checkout_lines(checkout_with_item)
     total = manager.calculate_checkout_total(
-        checkout_with_item, lines, address_usa_tx, discounts
+        checkout_with_item, list(checkout_with_item), discounts
     )
     total = quantize_price(total, total.currency)
     assert total == TaxedMoney(
@@ -306,7 +351,7 @@ def test_calculate_checkout_total_skip(
     skip_mock.return_value = True
     plugin_configuration()
     manager = get_plugins_manager()
-    manager.calculate_checkout_total(checkout_with_item, [], address_usa, [])
+    manager.calculate_checkout_total(checkout_with_item, [], [])
     skip_mock.assert_called_once
 
 
@@ -316,7 +361,7 @@ def test_calculate_checkout_total_invalid_checkout(
 ):
     plugin_configuration()
     manager = get_plugins_manager()
-    total = manager.calculate_checkout_total(checkout_with_item, [], address_usa, [])
+    total = manager.calculate_checkout_total(checkout_with_item, [], [])
     assert total == TaxedMoney(net=Money("0.00", "USD"), gross=Money("0.00", "USD"))
 
 
@@ -389,9 +434,11 @@ def test_calculate_checkout_total_excise_data(
     checkout.shipping_address = address_usa
     checkout.billing_address = address_usa
     shipping_method = shipping_zone.shipping_methods.get()
-    ShippingMethodChannelListing.objects.filter(shipping_method=shipping_method).update(
-        price_amount=0
-    )
+    # ShippingMethodChannelListing.objects.filter(shipping_method=shipping_method).update(
+    #     price_amount=0
+    # )
+    shipping_method.price_amount = 0
+    shipping_method.save()
     checkout.shipping_method = shipping_method
 
     metadata = {
@@ -408,9 +455,11 @@ def test_calculate_checkout_total_excise_data(
     variant.sku = variant_sku
     variant.private_metadata = metadata
     variant.save()
-    listing = variant.channel_listings.first()
-    listing.price_amount = Decimal(price)
-    listing.save()
+    # listing = variant.channel_listings.first()
+    # listing.price_amount = Decimal(price)
+    # listing.save()
+    variant.price_amount = Decimal(price)
+    variant.save()
     add_variant_to_checkout(checkout, variant, 1)
     checkout.save()
 
@@ -419,7 +468,7 @@ def test_calculate_checkout_total_excise_data(
     site_settings.save()
 
     lines = fetch_checkout_lines(checkout)
-    total = manager.calculate_checkout_total(checkout, lines, address_usa_tx, None)
+    total = manager.calculate_checkout_total(checkout, list(checkout), [])
     total = quantize_price(total, total.currency)
     assert total == TaxedMoney(
         net=Money(expected_net, "USD"), gross=Money(expected_gross, "USD")
@@ -448,14 +497,10 @@ def test_preprocess_order_creation(
     discounts = [discount_info]
     manager.preprocess_order_creation(checkout_with_item, discounts)
 
-
 @pytest.mark.vcr
 @override_settings(PLUGINS=["saleor.plugins.avatax.excise.plugin.AvataxExcisePlugin"])
 def test_preprocess_order_creation_wrong_data(
-    checkout_with_item,
-    address,
-    shipping_zone,
-    plugin_configuration,
+    checkout_with_item, address, shipping_zone, plugin_configuration,
 ):
     plugin_configuration()
 
@@ -473,12 +518,12 @@ def test_preprocess_order_creation_wrong_data(
 
 def test_api_post_request_handles_request_errors(product, monkeypatch):
     mocked_response = Mock(side_effect=RequestException())
-    monkeypatch.setattr("saleor.plugins.avatax.excise.requests.post", mocked_response)
+    monkeypatch.setattr(
+        "saleor.plugins.avatax.excise.utils.requests.post", mocked_response
+    )
 
     config = AvataxConfiguration(
-        username_or_account="test",
-        password_or_license="test",
-        use_sandbox=False,
+        username_or_account="test", password_or_license="test", use_sandbox=False,
     )
     url = "https://www.avatax.api.com/some-get-path"
 
@@ -490,12 +535,12 @@ def test_api_post_request_handles_request_errors(product, monkeypatch):
 
 def test_api_post_request_handles_json_errors(product, monkeypatch):
     mocked_response = Mock(side_effect=JSONDecodeError("", "", 0))
-    monkeypatch.setattr("saleor.plugins.avatax.excise.requests.post", mocked_response)
+    monkeypatch.setattr(
+        "saleor.plugins.avatax.excise.utils.requests.post", mocked_response
+    )
 
     config = AvataxConfiguration(
-        username_or_account="test",
-        password_or_license="test",
-        use_sandbox=False,
+        username_or_account="test", password_or_license="test", use_sandbox=False,
     )
     url = "https://www.avatax.api.com/some-get-path"
 
@@ -507,7 +552,9 @@ def test_api_post_request_handles_json_errors(product, monkeypatch):
 
 @pytest.mark.vcr
 @override_settings(PLUGINS=["saleor.plugins.avatax.excise.plugin.AvataxExcisePlugin"])
-def test_order_created_wrong_data(
+@patch("saleor.plugins.avatax.excise.plugin.api_post_request_task.delay")
+def test_order_created_calls_task(
+    api_post_request_task_mock,
     order_with_lines,
     address,
     address_usa_tx,
@@ -515,7 +562,7 @@ def test_order_created_wrong_data(
     shipping_zone,
     plugin_configuration,
 ):
-    plugin_configuration()
+    config = plugin_configuration()
     manager = get_plugins_manager()
 
     site_settings.company_address = address
@@ -525,11 +572,26 @@ def test_order_created_wrong_data(
     order_with_lines.shipping_method = shipping_zone.shipping_methods.get()
     order_with_lines.save()
 
-    with pytest.raises(TaxError) as e:
-        manager.order_created(order_with_lines)
-        # Fails due to skus not being configured in ATE
-    assert "does not exist or is not active" in e._excinfo[1].args[0]
 
+    # when
+    manager.order_created(order_with_lines)
+
+    # then 
+    transaction_url = "https://excisesbx.avalara.com/api/v1/AvaTaxExcise/transactions/create"
+    data = get_order_request_data(order_with_lines)
+    conf = {
+        data["name"]: data["value"] for data in config.configuration
+    }
+    configuration = {
+        'username_or_account': conf["Username or account"],
+        'password_or_license': conf["Password or license"], 'use_sandbox': True, 
+        'company_name': conf["Company name"], 
+        'autocommit': False
+    }
+
+    api_post_request_task_mock.assert_called_once_with(
+        transaction_url, asdict(data), configuration, order_with_lines.id
+    )
 
 @pytest.mark.vcr
 @override_settings(PLUGINS=["saleor.plugins.avatax.excise.plugin.AvataxExcisePlugin"])
@@ -551,9 +613,11 @@ def test_order_created(
     order_with_lines.shipping_address = address_usa_tx
     order_with_lines.shipping_method = shipping_zone.shipping_methods.get()
     shipping_method = shipping_zone.shipping_methods.get()
-    ShippingMethodChannelListing.objects.filter(shipping_method=shipping_method).update(
-        price_amount=0
-    )
+    # ShippingMethodChannelListing.objects.filter(shipping_method=shipping_method).update(
+    #     price_amount=0
+    # )
+    shipping_method.price_amount = 0
+    shipping_method.save()
     order_with_lines.shipping_method = shipping_method
 
     product.product_type = cigar_product_type
@@ -563,9 +627,11 @@ def test_order_created(
     variant.sku = "202015500"
     variant.save()
 
-    listing = variant.channel_listings.first()
-    listing.price_amount = Decimal(170)
-    listing.save()
+    # listing = variant.channel_listings.first()
+    # listing.price_amount = Decimal(170)
+    # listing.save()
+    variant.price_amount = Decimal(170)
+    variant.save()
 
     for order_line in order_with_lines.lines.all():
         order_line.product_name = product.name
